pyproject.toml
[tool.black]
line-length = 120
target-version = ["py39"]

[tool.pytest.ini_options]
log_cli = true
log_level = "debug"
testpaths = [
    "tests",
]

[tool.flake8]


[tool.poetry]
name = "py_skeleton"
version = "0.1.0"
description = "Modern up-to-date python project setup"
authors = ["Claudiu Stanciu <claudiu.a.stanciu@gmail.com>"]
license = "mit"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.12.1"
black = "^21.6b0"
flake8 = "^3.9.2"
pytest-asyncio = "^0.15.1"
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
